create general HowItWorks component
